Caching JWKs client with thread locking
